moved keybinding from assets to javascript
