+++ app/javascript/Components/Result/submission_selector.jsx
@@ -2,10 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {FilterModal} from "../Modals/filter_modal";
-import {
-  bind_keybindings,
-  unbind_all_keybindings,
-} from "../../../assets/javascripts/Results/keybinding";
+import {bind_keybindings, unbind_all_keybindings} from "./keybinding";
 import {ResultContext} from "./result_context";
 
 export class SubmissionSelector extends React.Component {
